Initial implementation of the game logic
The Brandi Class implements the base game. It keeps track of the
various elements required by the game.
These are:
 - Deck
 - Players and their Turns
 - Field

The Deck class stores the cards which have not yet been given to
players.

The Player class keeps track of information concerning the
player, such as:
 - Name
 - Color
 - Hand
 - The marbles in the players starting and end position

The Hand class keeps track of a players cards.

The Card class is used by the Deck and the Hand. It has a value
(2 - Ace), a color ('clubs', 'diamonds', 'hearts', 'spades'),
and a list of possible action by that card.

The Field class keeps track of the different marbles on the
playing field, but not the start or ending fields.