Redesign of field and marble class
The field is now a double linked list instead of a standard python
list. This facilitates marble movement along the field as well as
exiting the playing field to reach a players goal.

The marble object is now made to fit the field object and takes
care of remembering its position. It further remembers whether it
is allowed to enter the goal and wheteher it is blocking other
marbles in the starting position.