Brandi class now can do actions.
The class has been fitted to move marbles. This has been
implemented for starting actions, i.e. moving a marble out of ones
starting block, as well as for moving a 'normal' amount of steps
forward. The 'switch' action allowing the player to swap a marble
with one of the opponets marble and the seven individual steps for
a seven has not yet been implemented.

Furthermore a public state function has been implemented to
facilitate moving the game state to the client.